♻️ Extract PaymentInfo endpoint host into PaymentInfoEndpointHost
